Move Assertion into its own file
